requirements_dev.txt
Click>=8.1.3
docker>=6.1.3
factory-boy
Faker>=18.10.0
mysql-connector-python>=8.0.33
PyMySQL>=1.0.3
pytest>=7.3.1
pytest-cov
pytest-mock
pytest-timeout
pytimeparse2
simplejson>=3.19.1
types-simplejson
sqlalchemy>=2.0.0
sqlalchemy-utils
types-sqlalchemy-utils
tox
tqdm>=4.65.0
types-tqdm
packaging>=23.1
tabulate
types-tabulate
Unidecode>=1.3.6
typing_extensions
types-Pygments
types-colorama
types-mock
types-setuptools